Restructure v2 (#12)
* Restructure

* Added some comments

* Changed options system

* Bench command

* Fix to tree

* Uci print improvements

* First iteration pretty print

* Pretty print

* Format

* Clippy

* Clippy b2

+++ src/search/tree/tree_segment.rs
@@ -30,7 +30,7 @@ impl TreeSegment {
     }
 
     pub fn len(&self) -> usize {
-        self.length.load(Ordering::Relaxed) as usize
+        self.length.load(Ordering::Relaxed)
     }
 
     pub fn size(&self) -> usize {
@@ -51,7 +51,7 @@ impl TreeSegment {
         if new_index >= self.nodes.len() {
             return None;
         }
-        
+
         self.nodes[new_index].replace(state);
         Some(NodeIndex::from_parts(new_index as u32, self.segment))
     }
